Added remove users function
In this I had to change both the group users and the users GroupID if they were removed, which felt messy.
I've done it by creating a new user list for the group and overwriting the old one, then finding each of the users that was removed and deleting their group ID.
The database looks correct in the end but I think I could have done it with fewer database calls, or in a way that automatically removed the GroupID from the user when they were removed from the group through database traversal.